moves getUserWorkouts logic out of the server and into a WorkoutsController
