+++ server.js
@@ -1,6 +1,7 @@
 const LiftsController = require('./controllers/lifts-controller')
 const BodyareasController = require('./controllers/bodyareas-controller')
 const BodyareaLiftsController = require('./controllers/bodyarea-lifts-controller')
+const WorkoutsController = require('./controllers/workouts-controller')
 
 const express = require('express')
 const app = express()
@@ -75,42 +76,27 @@ app.delete('/api/v1/bodyareas/:bodyarea_id/lifts/:lift_id', (req, res) => {
 })
 
 app.get('/api/v1/users/:user_id/workouts', (req, res) => {
-  const id = req.params.user_id
-  // database.raw(`SELECT workouts.focus_area AS focus, workouts.workout_date AS date, lifts.name, sets.reps, sets.weight
-  //   FROM workouts
-  //   JOIN workout_lifts
-  //   ON workouts.id = workout_lifts.workout_id
-  //   JOIN lifts
-  //   ON workout_lifts.lift_id = lifts.id
-  //   JOIN sets
-  //   ON workout_lifts.id = sets.workout_lifts_id
-  //   WHERE workouts.user_id = ?
-  //   GROUP BY date, focus, name, reps, weight
-  //   ORDER BY date;`, [id])
+  WorkoutsController.getUserWorkouts(req, res)
+  // const id = req.params.user_id
+  // database.raw(`SELECT id, focus_area AS focus, workout_date AS date FROM workouts WHERE user_id = ?`, [id])
   // .then( data => {
+  //   const workouts = []
   //   data.rows.forEach(workout => {
-  //     console.log(workout)
+  //     const workoutId = workout.id
+  //     database.raw(`SELECT lifts.name, sets.reps, sets.weight FROM workout_lifts JOIN lifts ON workout_lifts.lift_id = lifts.id JOIN sets ON workout_lifts.id = sets.workout_lifts_id WHERE workout_lifts.id = ? ORDER BY workout_lifts.id`, [workoutId])
+  //     .then( singleWorkoutData => {
+  //       const lifts = []
+  //       singleWorkoutData.rows.forEach( aLift => {
+  //         lifts.push( { "name": aLift.name, "reps": aLift.reps, "weight": aLift.weight } )
+  //       })
+  //       return lifts
+  //     })
+  //     .then( lifts => {
+  //       const aSingleWorkout = { "id": workout.id, "date": workout.date, "focus": workout.focus, "lifts": lifts}
+  //       res.json(aSingleWorkout)
+  //     })
   //   })
   // })
-  database.raw(`SELECT id, focus_area AS focus, workout_date AS date FROM workouts WHERE user_id = ?`, [id])
-  .then( data => {
-    const workouts = []
-    data.rows.forEach(workout => {
-      const workoutId = workout.id
-      database.raw(`SELECT lifts.name, sets.reps, sets.weight FROM workout_lifts JOIN lifts ON workout_lifts.lift_id = lifts.id JOIN sets ON workout_lifts.id = sets.workout_lifts_id WHERE workout_lifts.id = ? ORDER BY workout_lifts.id`, [workoutId])
-      .then( singleWorkoutData => {
-        const lifts = []
-        singleWorkoutData.rows.forEach( aLift => {
-          lifts.push( { "name": aLift.name, "reps": aLift.reps, "weight": aLift.weight } )
-        })
-        return lifts
-      })
-      .then( lifts => {
-        const aSingleWorkout = { "id": workout.id, "date": workout.date, "focus": workout.focus, "lifts": lifts}
-        res.json(aSingleWorkout)
-      })
-    })
-  })
 })
 
 if (!module.parent) {
